+++ changes.md
@@ -25,6 +25,7 @@
 22. Fix - let correctly change marker attributes interactively 
 23. Fix - correctly use preloaded d3.js in notebooks/require.js environment  
 24. Use barwidth/baroffset parameters in lego plots
+25. Add fOptTitle, fOptLogx/y/z fields in JSROOT.gStyle
 
 
 ## Changes in 5.3.0